% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rxode2parseFunPtrs}
\alias{.rxode2parseFunPtrs}
\title{Return a list of the function parameters to hook into}
\usage{
.rxode2parseFunPtrs()
}
\value{
List of function pointers
}
\description{
Return a list of the function parameters to hook into
}
\examples{
.rxode2parseFunPtrs()
}
\author{
Matthew L. Fidler
}
\keyword{internal}
