% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linCmt.R
\name{.solve1pt}
\alias{.solve1pt}
\title{Solve 1 point from a compartmental model}
\usage{
.solve1pt(
  inp,
  t,
  k10,
  k12 = 0,
  k21 = 0,
  k13 = 0,
  k31 = 0,
  v = 1,
  rate = 0,
  ka = 0
)
}
\arguments{
\item{inp}{Input compartment amounts.  The first value represents
the depot compartment (if ka > 0) followed by the compartmetns
(1st, 2nd or 3rd) or the amount in the first compartment followed
by the next compartments in order (2nd, 3rd).}

\item{t}{The time after this point to sove for}

\item{k10}{elimination rate}

\item{k12}{rate from central to peripheral compartment}

\item{k21}{rate from peripheral to central compartment}

\item{k13}{rate from central to peripheral compartment #2}

\item{k31}{rate from peripheral compartment #2 to central}

\item{rate}{Rate that is active in the central compartment}

\item{ka}{The oral absorption rate}
}
\value{
All compartment values solved to t after current value
}
\description{
Solve 1 point from a compartmental model
}
\author{
Matthew L. Fidler
}
\keyword{internal}
