% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudf.R
\name{.udfEnvLockIfExists}
\alias{.udfEnvLockIfExists}
\title{Lock the UDF function if the object exits inside of it}
\usage{
.udfEnvLockIfExists(obj)
}
\arguments{
\item{obj}{object to check to see if it exists}
}
\value{
nothing called for side effects
}
\description{
Lock the UDF function if the object exits inside of it
}
\author{
Matthew L. Fidler
}
\keyword{internal}
