% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudf.R
\name{.udfFindAndLock}
\alias{.udfFindAndLock}
\title{Find an object and if exists lock the environment}
\usage{
.udfFindAndLock(object, envirList = list())
}
\arguments{
\item{object}{is the R object to find}

\item{envirList}{a list of enviroments search}
}
\value{
This returns if the environment has been locked
}
\description{
Find an object and if exists lock the environment
}
\author{
Matthew L. Fidler
}
\keyword{internal}
