% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linCmt.R
\name{linCmt}
\alias{linCmt}
\title{R implementation of \code{rxode2parse}/\code{rxode2} \code{linCmt()} solutions}
\usage{
linCmt(
  data,
  ...,
  fdepot = 1,
  fcentral = 1,
  ratedepot = 0,
  ratecentral = 0,
  durdepot = 0,
  durcentral = 0,
  lagdepot = 0,
  lagcentral = 0,
  addlKeepsCov = FALSE,
  addlDropSs = TRUE,
  ssAtDoseTime = TRUE,
  scale = 1,
  gradient = FALSE,
  keep = NULL,
  covsInterpolation = c("locf", "linear", "nocb", "midpoint"),
  sc = 1,
  sm = 1/sc
)
}
\arguments{
\item{data}{input dataset, may have some parameters in the dataset
itself; It also uses the standard \code{rxode2} dataset format which
is translated as usual with \code{etTrans()}.  This procedure will
produce the underlying \code{rxode2} model code and variables to
translate the compartmental model}

\item{...}{Parameters that can be passed/parsed and integrated into
the \code{data} for a linear compartment solution.  You can either
have the parameters in the dataset itself or in this parameter
block.  The type of linear compartmental model is determined by
the parameters you specify in either the \code{data} or this extra set
of arguments.}

\item{fdepot}{The bioavailibility of doses to the depot.  Also
affects rate/dur. The default is \code{1} or there is no
bioavailibilty effect.  In \code{rxode2} you can specify this in a
\code{linCmt()} model by adding \verb{f(depot)=}.}

\item{fcentral}{The bioavailibility of doses to the central compartment.  Also
affects rate/dur. The default is \code{1} or there is no
bioavailibilty effect.  In \code{rxode2} you can specify this in a
\code{linCmt()} model by adding \verb{f(central)=}.}

\item{ratedepot}{The rate of doses that are modeled with \code{rate=-1}
when dosed to the \code{depot} compartment.  This can be added to a
\code{rxode2} \code{linCmt()} model by adding \verb{rate(depot)=} to the model.}

\item{ratecentral}{The rate of doses that are modeled with
\code{rate=-1} when dosed to the \code{central} compartment.  This can be
added to a \code{rxode2} \code{linCmt()} model by adding \verb{rate(central)=}.}

\item{durdepot}{The duration of infusion in the depot compartment
with doses that are modeled with \code{rate=-2}.  This can be added to
a \code{rxode2} \code{linCmt()} model by adding \verb{dur(depot)=} to the model.}

\item{durcentral}{The duration of infusions in the central
compartment with doses that are modeled with \code{rate=-2}. This can
be added to a \code{rxode2} \code{linCmt()} model by adding \verb{dur(central)=}
to the model.}

\item{lagdepot}{The lag time of the depot compartment.  Can be
specified in an \code{rxode2} model by \verb{alag(depot)=}}

\item{lagcentral}{he lag time of the depot compartment.  Can be
specified in an \code{rxode2} model by \verb{alag(central)=}}

\item{addlKeepsCov}{This determines if the additional dosing items
repeats the dose only (\code{FALSE}) or keeps the covariates at the
record of the dose (\code{TRUE})}

\item{addlDropSs}{When there are steady state doses with an \code{addl}
specification the steady state flag is dropped with repeated
doses (when \code{TRUE}) or retained (when \code{FALSE})}

\item{ssAtDoseTime}{Boolean that when \code{TRUE} back calculates the
steady concentration at the actual time of dose, otherwise when
\code{FALSE} the doses are shifted}

\item{scale}{Scaling factor for the final \code{Cc} output}

\item{gradient}{If \code{TRUE}, this will add the gradients for the
linear compartment models to the output, otherwise it only shows
the concentration in the central compartment}

\item{keep}{This is a named vector of items you want to keep in the final rxode2 dataset.
For added rxode2 event records (if seen), last observation carried forward will be used.}

\item{sc}{scaling parameter of the central compartment like
NONMEM's scaling parameter. The concentration is divided by this
parameter value.}

\item{sm}{multiplicative scaling parameter.  Unlike the NONMEM
scaling parameter, the concentration is multiplied by this
parameter.}
}
\value{
A dataframe containing the linear compartment solution of
the model appended as a column to the input data as \code{Cc}
}
\description{
This interfaces the C/C++ linCmt() interface for use in R without
having to compile any \code{rxode2} models
}
\examples{

d <- nlmixr2data::nmtest

names(d) <- sub("lagt", "lagcentral",
            sub("bioav", "fdepot",
            sub("rat2", "ratecentral",
            sub("dur2", "durcentral", names(d)))))

ret <- linCmt(d, cl=1.1, v=20, ka=1.5)

}
\author{
Matthew L. Fidler
}
