% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rxLinCmt}
\alias{rxLinCmt}
\title{Linear compartmental concentrations based on input data}
\usage{
rxLinCmt(
  data,
  ...,
  tlag = 0,
  F = 1,
  rate1 = 0,
  dur1 = 0,
  tlag2 = 0,
  F2 = 1,
  rate2 = 0,
  dur2 = 0,
  verbose = FALSE
)
}
\arguments{
\item{data}{This dataest should have the following columns:
\itemize{
\item \code{evid} - \code{rxode2} style evid
\item \code{amt} - amount of dose
\item \code{time} - time of observation
}

The following columns are optional:
\itemize{
\item \code{id} the identifier of the individual
\item \code{cmt} the compartment of dose
\item \code{rate} the rate of dose
\item \code{dur} the duration of the dose
\item \code{ii} the inter-dose interval
\item \code{ss} the steady state flag
}

This cannot have the an \code{id} column}

\item{...}{Parameters for linear compartment model, parsed automatically}

\item{tlag}{This is the lag time of the first compartment.}

\item{F}{This is the bioavailability of the first compartment}

\item{rate1}{This is the modeled rate for the first compartment}

\item{dur1}{This is the modeled duration for the first compartment}

\item{tlag2}{This is the lag time of the second compartment}

\item{F2}{This is the modeled bioavailability of the second
compartment}

\item{rate2}{This is the modeled rate of the second compartment}

\item{dur2}{This is the modeled duration of the second compartment}

\item{verbose}{boolean that when TRUE provides a message about the detected pk parameters
and the detected compartmental model.  By default this is \code{FALSE}.}
}
\value{
original dataset \code{data} with an additional column \code{Cp} with
the
}
\description{
Linear compartmental concentrations based on input data
}
\examples{

rxLinCmt(data=nlmixr2data::theo_sd, CL = 29.4, V = 3)

}
\author{
Matthew L. Fidler
}
