% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tran.R
\name{rxode2parseMd5}
\alias{rxode2parseMd5}
\title{Get the MD5 hash of the current language revision}
\usage{
rxode2parseMd5()
}
\value{
md5 hash of language revision
}
\description{
Get the MD5 hash of the current language revision
}
\examples{
rxode2parseMd5()
}
\author{
Matthew L. Fidler
}
