% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tran.R
\name{rxode2parse}
\alias{rxode2parse}
\title{Internal translation to get model variables list}
\usage{
rxode2parse(model)
}
\arguments{
\item{model}{Model (either file name or string)}
}
\value{
A rxModelVars object that has the model variables of a rxode2 syntax expression
}
\description{
Internal translation to get model variables list
}
\examples{
rxode2parse("a=3")
}
